Move specs parsing functionality to a new pkg/spec package
