+++ cmd/fleetctl/preview.go
@@ -21,6 +21,7 @@ import (
 	"github.com/fleetdm/fleet/v4/orbit/pkg/update"
 	"github.com/fleetdm/fleet/v4/pkg/fleethttp"
 	"github.com/fleetdm/fleet/v4/pkg/open"
+	"github.com/fleetdm/fleet/v4/pkg/spec"
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/service"
 	"github.com/mitchellh/go-ps"
@@ -297,7 +298,14 @@ Use the stop and reset subcommands to manage the server and dependencies once st
 				}
 			}
 
-			err = applyYamlBytes(c, buf, client)
+			specs, err := spec.GroupFromBytes(buf)
+			if err != nil {
+				return err
+			}
+			logf := func(format string, a ...interface{}) {
+				fmt.Fprintf(c.App.Writer, format, a...)
+			}
+			err = client.ApplyGroup(c.Context, specs, logf)
 			if err != nil {
 				return err
 			}
